fixed c4172 error

+++ UnitTest/enem.cpp
@@ -19,8 +19,8 @@ enem::~enem()
 
 void enem::draw()
 {
-	Rectangle source = { 0, 0, texture.width - 1, texture.height - 1};
-	Rectangle dest = {pos.x, pos.y, texture.width, texture.height };
+	Rectangle source = { 0.0f, 0.0f, texture.width - 1.0f, texture.height - 1.0f};
+	Rectangle dest = {(float)pos.x, (float)pos.y, (float)texture.width, (float)texture.height };
 	Vector2 center = { source.width / 2, source.height / 2 };
 	// texture
 	DrawTexturePro(texture, source, dest, center, forwardDir.angleBetween(vec2(0, 1)), WHITE); // make rot rad
